Refactor CMS Signal Handling
https://openedx.atlassian.net/browse/EDUCATOR-389
