+++ cms/djangoapps/contentstore/tests/test_proctoring.py
@@ -9,7 +9,7 @@ from pytz import UTC
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory, ItemFactory
 
-from contentstore.signals import listen_for_course_publish
+from contentstore.signals.handlers import listen_for_course_publish
 
 from edx_proctoring.api import (
     get_all_exams_for_course,
